update of entity

+++ project/src/main/java/com/example/project/dao/ExerciseRepository.java
@@ -3,5 +3,5 @@ package com.example.project.dao;
 import com.example.project.entity.Exercise;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExerciseRepository extends JpaRepository<Exercise,Integer> {
+public interface ExerciseRepository extends JpaRepository<Exercise,Long> {
 }
